refactor: Update package.json version and add @webscopeio/react-textarea-autocomplete dependency
This commit updates the version in package.json from 0.1.4 to 0.1.5 and adds the @webscopeio/react-textarea-autocomplete dependency to the project. This dependency will be used for implementing autocomplete functionality in the ConnectionApp component.

Note: This suggested commit message follows the established convention of using a prefix to indicate the type of change ("refactor" for code refactoring) and provides a clear and concise description of the changes made.