refactor: Update package.json version and dependencies
This commit updates the package.json version from 0.1.6 to 0.1.7 and updates the dependencies in the project. Specifically, the providers array in the getQuotes function in quotes.js is commented out. This change suggests that the MAYACHAIN provider is no longer used. Additionally, the corsProxy value in IRCWindow.jsx is updated to 'http://poxy.winbit32.com:8000/'. These updates ensure that the project is using the latest stable versions of the dependencies and make necessary adjustments to the code.

Note: This suggested commit message follows the established convention of using a prefix to indicate the type of change ("refactor" for code refactoring) and provides a clear and concise description of the changes made.