chore: Remove create-hash dependency from package.json
This commit removes the create-hash dependency from the package.json file. The dependency was previously added in version 0.1.14, but it is no longer needed. This change helps to streamline the dependencies of the Winbit32 application.

Note: This suggested commit message follows the established convention of using a prefix to indicate the type of change ("chore" for general maintenance) and provides a clear and concise description of the changes made.