refactor: Update getAssetValue function in quote.js
This commit updates the getAssetValue function in the quote.js file. The changes include adding a new condition to set the address value for the assetValue object when the symbol is 'XRD' and the chainId is "radix-mainnet". This ensures that the correct address is assigned for the XRD token on the Radix mainnet. The modification improves the accuracy of asset value calculations in the Winbit32 application.

Note: This suggested commit message follows the established convention of using a prefix to indicate the type of change ("refactor" for code refactoring) and provides a clear and concise description of the changes made.