refactor: Update CSS to float the icon in the welcome dialog
This commit updates the win3.css file to float the icon in the welcome dialog to the left. The addition of the 'float: left' property to the '.welcome-dialog > .content > div.icon' selector ensures that the icon is aligned to the left side of the dialog. This change improves the visual layout of the welcome dialog in the Winbit32 application.